requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aioquic==1.2.0
aioredis==1.3.1
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.4.0
audioop-lts==0.2.1
backoff==2.2.1
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
browserhistory==0.1.2
cachelib==0.13.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
cfonts==1.5.2
charset-normalizer==3.4.4
circuitbreaker==2.1.3
click==8.2.1
cloudpathlib==0.21.0
clr_loader==0.2.7.post0
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
crypto==1.4.1
cryptography==45.0.7
customtkinter==5.2.2
cymem==2.0.11
Cython==3.1.3
darkdetect==0.8.0
decorator==5.2.1
discord-interactions==0.4.0
discord.py==2.6.3
DLL-Injector==1.0.5
dnspython==2.8.0
edge-tts==7.0.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.119.0
fastapi-cache==0.1.0
ffmpeg==1.4
filelock==3.18.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Cors==5.0.0
frida==17.2.11
frida-tools==14.4.1
frozenlist==1.8.0
fsspec==2025.3.2
future==1.0.0
geocoder==1.38.1
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.1
h11==0.16.0
h2==4.3.0
hiredis==3.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
kaitaistruct==0.10
keyauth==1.0.3
keyboard==0.13.5
langcodes==3.5.0
language_data==1.3.0
ldap3==2.9.1
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mitmproxy==12.1.2
mitmproxy-windows==0.12.7
mitmproxy_rs==0.12.7
more-itertools==10.7.0
mouse==0.7.1
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.1
mss==10.0.0
multidict==6.7.0
murmurhash==1.0.12
Naked==0.1.32
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20250625
packaging==25.0
passlib==1.7.4
pefile==2023.2.7
pillow==11.3.0
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.30.0
protobuf-decoder==0.4.0
psutil==7.1.0
publicsuffix2==2.20191221
pvporcupine==3.0.5
pyarmor==9.1.7
pyarmor.cli.core==7.6.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.3
pydantic_core==2.33.1
pydivert==2.1.0
pygame==2.6.1
PyGetWindow==0.0.9
PyGithub==2.8.1
Pygments==2.19.1
pyinjector==1.1.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.7
PyJWT==2.10.1
pylsqpack==0.3.22
Pymem==1.14.0
pymongo==4.15.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PyRect==0.2.0
PyScreeze==1.0.1
pystyle==2.9
python-cfonts==1.5.2
python-dotenv==1.0.1
python-multipart==0.0.20
pythonnet==3.0.5
pyttsx3==2.99
pytube==15.0.0
pytweening==1.2.0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.2
ratelim==0.1.6
redis==7.0.1
regex==2024.11.6
requests==2.32.5
rich==14.0.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scapy==2.5.0
scipy==1.16.1
service-identity==24.2.0
setuptools==80.9.0
shellescape==3.8.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.40
srsly==2.5.1
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.48.0
sv_ttk==2.6.1
sympy==1.13.3
tabulate==0.9.0
termcolor==3.0.1
thinc==8.3.6
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
urwid==2.6.16
uvicorn==0.37.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==13.1
Werkzeug==3.1.3
win10toast==0.9
wrapt==1.17.2
wsproto==1.2.0
yarl==1.22.0
yt-dlp==2025.9.26
zstandard==0.24.0